Add another test for two simultaneous ECU's maintaining their own state

+++ src/domain/ECU.cpp
@@ -1,7 +1,7 @@
 #include "domain/ECU.h"
 #include <string_view>
 
-ECU::ECU(std::string_view ecuName, std::uint8_t diagnosticRequestCANId, std::uint8_t diagnosticResponseCANId) 
+ECU::ECU(std::string_view ecuName, std::uint16_t diagnosticRequestCANId, std::uint16_t diagnosticResponseCANId) 
     : m_ecuName {ecuName}
     , m_diagnosticRequestCANId {diagnosticRequestCANId}
     , m_diagnosticResponseCANId {diagnosticResponseCANId} {}
